+++ System.Drawing/FontFamily-CoreText.cs
@@ -185,9 +185,9 @@ namespace System.Drawing
 				return (int)Math.Round(font.DescentMetric / font.Size * font.UnitsPerEmMetric);
 			case  Metric.LineSpacing:
 				float lineHeight = 0;
-				lineHeight += font.AscentMetric;
-				lineHeight += font.DescentMetric;
-				lineHeight += font.LeadingMetric;
+				lineHeight += (float)font.AscentMetric;
+				lineHeight += (float)font.DescentMetric;
+				lineHeight += (float)font.LeadingMetric;
 				return (int)Math.Round(lineHeight / font.Size * font.UnitsPerEmMetric);
 			}
