+++ System.Drawing/Image.cs
@@ -92,7 +92,7 @@ namespace System.Drawing {
 		public int Height {
 			get {
 				var b = this as Bitmap;
-				return b == null ? 0 : b.NativeCGImage.Height;
+				return b == null ? 0 : (int)b.NativeCGImage.Height;
 			}
 		}
 		
@@ -117,7 +117,7 @@ namespace System.Drawing {
 		public int Width {
 			get {
 				var b = this as Bitmap;
-				return b == null ? 0 : b.NativeCGImage.Width;
+				return b == null ? 0 : (int)b.NativeCGImage.Width;
 			}
 		}
